Struct binjs_shared::ast::Path

pub struct Path<I, F> where
    I: Debug + PartialEq,
    F: Debug + PartialEq,  { /* fields omitted */ }

The path followed when walking an AST.

Designed to be used both to quickly find out how to contextually handle a specific node and for error-reporting.

use binjs_shared::ast::Path;

let mut path = Path::new();
assert!(path.get(0).is_none());

// Once we have entered both an interface and a field, `path.get(0)` will be `Some`.
path.enter_interface("Interface 1");
assert!(path.get(0).is_none());

path.enter_field("Field 1");

{
  let item = path.get(0).unwrap();
  assert_eq!(item.field, "Field 1");
  assert_eq!(item.interface, "Interface 1");
}

path.enter_interface("Interface 2");
path.enter_field("Field 2");

{
  let item = path.get(0).unwrap();
  assert_eq!(item.field, "Field 2");
  assert_eq!(item.interface, "Interface 2");
}
{
  let item = path.get(1).unwrap();
  assert_eq!(item.field, "Field 1");
  assert_eq!(item.interface, "Interface 1");
}

// We need to exit the field before exiting the interface.
path.exit_field("Field 2"); // Exiting the wrong field would panic.
path.exit_interface("Interface 2"); // Exiting the wrong interface would panic.
path.exit_field("Field 1"); // Exiting the wrong field would panic.
path.exit_interface("Interface 1"); // Exiting the wrong interface would panic.

Methods

impl<I, F> Path<I, F> where
I: Debug + PartialEq,
F: Debug + PartialEq,

pub fn new() -> Self

Create an empty Path.

pub fn enter_interface(&mut self, node: I)

Enter an interface.

All calls to enter_interface MUST be balanced with calls to exit_interface.

pub fn exit_interface(&mut self, node: I)

pub fn enter_field(&mut self, field: F)

pub fn exit_field(&mut self, field: F)

pub fn len(&self) -> usize

pub fn get(&self, index: usize) -> Option<&PathItem<I, F>>